refactor : OAuth2 코드 리팩토링 2차 수정
