+++ build.gradle
@@ -31,7 +31,13 @@ dependencies {
     annotationProcessor 'org.projectlombok:lombok'
 
     // Spring Security
-//    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+
+    // JWT 관련 기능
+    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 
     // https://mvnrepository.com/artifact/net.coobird/thumbnailator / 썸네일
     implementation group: 'net.coobird', name: 'thumbnailator', version: '0.4.19'
