+++ src/main/java/org/gdg/zipte_gdg/api/service/oauth/response/NaverResponse.java
@@ -1,4 +1,4 @@
-package org.gdg.zipte_gdg.api.service.member.response.oauth2;
+package org.gdg.zipte_gdg.api.service.oauth.response;
 
 import java.util.Map;
 import java.util.Optional;
