+++ src/main/java/org/gdg/zipte_gdg/domain/member/MemberRepository.java
@@ -11,5 +11,5 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     Member findByUsername(String username);
 
     @Query("select m from Member m where m.email =:email")
-    Optional<Member> findByEmail(@Param("email") String email);
+    Member findByEmail(@Param("email") String email);
 }
